+++ src/net/thornydev/mybatis/koan/koan08/Koan08.java
@@ -70,7 +70,7 @@ public class Koan08 {
 			assertEquals(1, n);
 			
 			// validate that the insert worked
-			int totalCountries = session.selectOne("getCountryCount");
+			int totalCountries = (Integer) session.selectOne("getCountryCount");
 			assertEquals(110, totalCountries);
 			
 			Country southSudan = session.selectOne("getCountryById", 1000);
@@ -109,7 +109,7 @@ public class Koan08 {
 			int n = -1;
 			assertEquals(1, n);
 			
-			int totalCountries = session.selectOne("getCountryCount");
+			int totalCountries = (Integer) session.selectOne("getCountryCount");
 			assertEquals(109, totalCountries);
 			
 			Country northSudan = session.selectOne("getCountryById", 89);
@@ -147,7 +147,7 @@ public class Koan08 {
 			n = session.insert("insertCountry", c1001);
 			assertEquals(1, n);
 			
-			int totalCountries = session.selectOne("getCountryCount");			
+			int totalCountries = (Integer) session.selectOne("getCountryCount");			
 			assertEquals(111, totalCountries);
 			
 			// second delete it via mapper class
@@ -155,14 +155,14 @@ public class Koan08 {
 			
 			n = mapper.deleteCountryById( c1000.getId() );
 			assertEquals(1, n);
-			totalCountries = session.selectOne("getCountryCount");			
+			totalCountries = (Integer) session.selectOne("getCountryCount");			
 			assertEquals(110, totalCountries);
 			
 			// TODO: delete record with id 1001 by using the whole Country
 			//       object rather than just passing the id
 			n = -1;
 			assertEquals(1, n);
-			totalCountries = session.selectOne("getCountryCount");			
+			totalCountries = (Integer) session.selectOne("getCountryCount");			
 			assertEquals(109, totalCountries);
 			
 		} finally {
@@ -185,7 +185,7 @@ public class Koan08 {
 		try {
 			session = sessionFactory.openSession();
 			// first get max country id and cache it for assert testing later
-			int maxId = session.selectOne("getMaxCountryId");
+			int maxId = (Integer) session.selectOne("getMaxCountryId");
 			
 			// TODO: create the Country "South Sudan", but leave the id blank
 			Country c = null;
@@ -194,7 +194,7 @@ public class Koan08 {
 			assertEquals(1, n);
 			assertEquals(maxId + 1, c.getId());  // id should have been filled in and set to new maxId
 			
-			int totalCountries = session.selectOne("getCountryCount");
+			int totalCountries = (Integer) session.selectOne("getCountryCount");
 			assertEquals(110, totalCountries);
 			
 			Country southSudan = session.selectOne("getCountryById", maxId + 1);
